capteurs/requirements.txt
paho-mqtt
numpy